add exception basic function
