+++ lab3/exception.c
@@ -1,0 +1,7 @@
+#include "uart.h"
+
+void
+el2_sync_message(unsigned long elr_el2, unsigned long esr_el2)
+{
+    // todo
+}
