+++ lab3/exception.h
@@ -1,0 +1,4 @@
+#ifndef _EXCEPTION_H
+#define _EXCEPTION_H
+
+#endif
