+++ lab3/loadimg.h
@@ -1,0 +1,6 @@
+#ifndef _LOADIMG_H
+#define _LOADIMG_H
+
+void load_image(char *input_address);
+
+#endif
