--- lab3/ut.h
@@ -1,1 +1,0 @@
-void branchAddr(void *addr);
