--- lab3/ut.s
@@ -1,3 +1,0 @@
-.global branchAddr
-branchAddr:
-    br x0
